requirements.txt
blinker==1.4
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
Flask-FlatPages==0.7.2
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.4.4
Flask-Themes==0.1.3
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.3
Markdown==3.3.3
MarkupSafe==1.1.1
mimerender==0.6.0
pymongo==3.11.3
python-dateutil==2.8.1
python-mimeparse==1.6.0
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.3.23
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
flask_profiler==1.8.1
email_validator
kafka-python
